Scraper/requirements.txt
Django
python-dotenv
djangorestframework
drf-yasg
markdown       
mysqlclient
Pillow
requests
django-filter
